+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for QR Manager
 
+## 5.0.3 - 2024-10-08
+- Added redirect count to route index view
+- Added redirect count to route edit page
+- Added validation to prevent two routes being created with the same Entry URI
+
 ## 5.0.2.1 - 2024-09-20
 - Fixed composer version/release mismatch